Convert char generators to codepoint generators
We were essentially generating codepoints anyway, so this eliminates the back and forth conversions, speeding things up and making the intention clear.  The downside is that there isn't a nice way to generate multicodepoint characters, but the previous implementation didn't support this cleanly either.  This change better aligns the types with the actual intention.